+++ dipimage/dt.m
@@ -4,18 +4,18 @@
 %  image_out = dt(image_in,edgeCondition,method)
 %
 % PARAMETERS:
-%  edgeCondition: the value of pixels outside the image bounds,
-%      can be 'background' or 'object', or equivalently 0 or 1.
-%  method: 'fast', 'ties', 'true', 'brute force'
+%  edgeCondition: the value of pixels outside the image bounds, can be
+%      'background' or 'object', or equivalently 0 or 1.
+%  method: 'separable', 'fast', 'ties', 'true', 'brute force', or 'square'
 %
 % DEFAULTS:
 %  edgeCondition = 'object'
-%  method = 'fast'
+%  method = 'separable'
 %
 % DIPlib:
 %  This function calls the DIPlib function dip::EuclideanDistanceTransform
 
-% (c)2017, Cris Luengo.
+% (c)2017-2018, Cris Luengo.
 % Based on original DIPlib code: (c)1995-2014, Delft University of Technology.
 % Based on original DIPimage code: (c)1999-2014, Delft University of Technology.
 %